Ignore receivers with bad clocks
clock sync is checked every 30 seconds, if a receiver has a bad sync
state at that time, it gets 1 added to his bad_sync score, this score is
capped at 6.
If the clock sync is good when checked, 0.05 is subtracted from the
bad_sync score.
1 bad observed sync state will take 20 good sync states to get back on
our good side, which takes 10 minutes.
If the bad_sync score is maxed out, it will take 60 minutes to no longer
be ignored for position calculation.