Ignore receivers with bad clocks
clock sync is checked every 30 seconds, if a receiver has a bad sync
state at that time, it gets 0.2 to 1 added to his bad_sync score.
The bad_sync score is capped at 6, this is equivalent to a 30 minute
timeout.
If the clock sync is good when checked, 0.1 is subtracted from the
bad_sync score.
A bad observed sync state will take 2 to 10 good sync states to get back
on our good side, which takes 1 to 5 minutes, depending how bad the sync
state was.